// src/app/component/ubs/models/personalData.model.ts
export interface PersonalData {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  addressComment: string;
  city: string;
  district: string;
  street: string;
  houseCorpus: string;
  entranceNumber: string;
  houseNumber: string;
  longitude?: number;
  latitude?: number;
}